Optimize page load speeds by generating code in content scripts
This commit removes final injection code generation from background scripts and moves it into content scripts.
This is because most of the time, the configuration including injection code had to be sent from background
scripts using syncMessage, which, for requests of that size (~700kB for Recommended level with FPD), took
more than 150ms. In comparison, the process of code generation including wrapper definition loading didn't
exceed 30ms. All measurements were made on Chrome. Because wrapping occurs during early injection on Firefox,
the time savings are not as large, but still apparent (the syncMessage request has to always finish, even
when wrapping is actually completed by the time it does, content script and page load is paused until then).

* Move necessary files between content and page scripts
* Move FPD level initialization in fp_code_builders.js to new background script fp_levels.js
* Optimize injection code size (don't generate xray wrapper when in Chrome, move property descriptor search into WrapHelper)
* Injection code wrapping is now called dynamicaly in document_start.js, remove it from levels.js and level_cache.js
* getContentConfiguration now waits for FPD level initialization as well as JSS level initialization
* Reduce size of page configuration for faster transfer and injection times by sending just what's neccessary and using arrays instead of dicts
* No changes to injection results, protection is still guaranteed