Add optimized farbling WebAssembly implementation
This commits adds a new implementation of image and audio farbling
functions written in AssemblyScript. The farbling results must stay
consistent with the JS version as the WASM module injection is not
completely reliable and the JS version might have to be used as a
fallback. For Chromium, a onHeadersReceived handler to modify received
CSP headers must be registered because WASM initialization is subject to
page CSP rules, the module must also be made available as a web
accessible resource. The compiler uses NodeJS and has been added as a step
during the build process using the provided makefile. This makes node a
requirement for building JShelter.

The WASM module is initialized on wrapping start just after WrapHelper
and it's functions are exposed to wrappers in the wasm object. For
storing and retrieving data from wasm memory, the included getter and
setter must be used (to correctly work around xray protection on
Firefox). Before any farbling operations begin, wasm.ready must be
checked if the module is correctly initialized and the size of memory must
be checked and expanded if needed by calling wasm.memory_grow and specifying
the total bytes neccesarry for the current operation.